feat: move shared util code into core
