+++ core/contract.ts
