+++ core/index.ts
@@ -1,5 +1,11 @@
 // @stellar-scaffold/ui-core — shared, framework-agnostic layer for the UI monorepo.
-// Populated by the core audit: runtime util (wallet/storage/friendbot/contract
-// helpers), framework-agnostic business logic, shared base config, base styles.
-// Excludes the framework-specific Provider Layer and UI Component Layer.
-export {};
+// One authored copy, imported by every template (and the generated Contract
+// Clients in core/clients). Framework-specific code (providers, stores, UI
+// components) stays in each template.
+
+export * from "./env" // rpcUrl, networkPassphrase, stellarNetwork, horizonUrl, network, labPrefix
+export * from "./contract" // shortenContractId
+export * from "./friendbot" // getFriendbotUrl
+export * from "./subscription" // subscribeToEvents
+export * from "./wallet" // connectWallet, disconnectWallet, fetchBalances, wallet, MappedBalances
+export { default as storage } from "./storage"
