+++ core/package.json
@@ -5,5 +5,10 @@
   "type": "module",
   "exports": {
     ".": "./index.ts"
+  },
+  "dependencies": {
+    "@creit.tech/stellar-wallets-kit": "^1.9.5",
+    "@stellar/stellar-sdk": "^15.1.0",
+    "zod": "^4.3.5"
   }
 }
