+++ core/storage.ts
