+++ templates/react/src/components/WalletButton.tsx
@@ -1,10 +1,9 @@
+import { connectWallet, disconnectWallet } from "@stellar-scaffold/ui-core"
 import { useRef } from "react"
 import { useWallet } from "../hooks/useWallet"
-import { connectWallet, disconnectWallet } from "../util/wallet"
 import styles from "./WalletButton.module.css"
 
-const shortAddress = (addr: string) =>
-	`${addr.slice(0, 4)}...${addr.slice(-4)}`
+const shortAddress = (addr: string) => `${addr.slice(0, 4)}...${addr.slice(-4)}`
 
 export const WalletButton = () => {
 	const dialogRef = useRef<HTMLDialogElement>(null)
